requirements.txt
# Python dependencies for Airflow ETL pipeline
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.0
# Removed postal package due to complex libpostal dependency - using simple address normalization instead

# OSM Collection DAG dependencies
shapely>=2.0.0
geopandas>=0.12.0
geopy>=2.3.0
